requirements.txt
python-telegram-bot==13.7
transformers==4.21.0
requests==2.28.1
python-dotenv==0.20.0